BAU: Extract SqsTestHelper
